_src/requirements.txt
flask
requests
beautifulsoup4
lxml
python-dotenv
aiohttp
asyncio
pydantic
agent-framework
agent-framework-core
azure-identity
# agent-framework[viz]
# agent-framework --pre